+++ changelog.md
@@ -2,3 +2,7 @@
 
 ## Version 1.0.0
 - Init module
+
+## Version 1.0.1
+- Sort conveniences in offer
+- Toast messages